api/pyproject.toml
[project]
name = "api-panel-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "autoflake>=2.3.1",
    "bcrypt>=4.3.0",
    "black>=25.1.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "openai>=1.82.0",
    "passlib>=1.7.4",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.11.3",
    "python-jose>=3.4.0",
    "python-json-logger>=3.3.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "uvicorn[standard]>=0.34.2",
]
